NewMCTS: check if each inital action is comparable
the old implementation was flawed. It checked if the Action
type was comparable which always returned true.

This change checks the base type of each action returned
from the inital state, and checks if each type is comparable.

The reason we check each separate action is that there's no
reason that the slice won't contain multiple base types.