remove isParentOf func
The reason this function existed was to filter duplicate moves from
the same game state. The reason for its removal is that the filtering
of moves should be moved onto the user of this package. 9 times out of
10, there would be no moves to filter, so it just wastes cpu cycles on
searching for a non-existent duplicate through the node's children. This
loop would occur each time a duplicate game state was found, wasting
time for games where duplicate states are common.