add Hash() as a required method on Game
The game state is now allowed to be uncomparable, as we use the hash,
assumed to be comparable, to map game states to nodes. This might seem,
like kicking the rock down the road, and it is to an extent, but this
allows the underlying game state to hold slices and maps.

This commit is harmless to current implemenations as the Hash function
can return the game state itself.