rework action selection system
We offload the action selection to the user of this package.
Theoritically, gmcts only needs to know how many actions there are,
and pick random actions from there.

This change is a rather massive change for current implementations,
but significant speed gains are available due to not having to convert
actions from values to interfaces back to values.

This change also removes the need for an Action interface, as the package
no longer needs to hold a list of random actions.